Fix bug in WMS client layers where PROCESSING "RESAMPLE=..." is given.
The bug occurs when a PROCESSING "RESAMPLE=" directive is given and 
there is not a re-projection between the WMS server and the MapServer
output. The bug occurs because the resampling code expects a valid
world file, however MapServer only writes a world file if
re-projection is necessary. This change writes a world file if either
a re-projection is necessary or resampling is specified.


git-svn-id: http://svn.osgeo.org/mapserver/trunk@9839 7532c77e-422f-0410-93f4-f0b67bdd69e2