change linking against libfastcgi (#4296)
fastcgi functions are only called by the mapserv binary,
so the fastcgi headers and library only needs to be
referenced when compiling/linking mapserv.c.

this patch moves the fastcgi io handlers from mapio.c to
mapserv.c, and changes the compilation flags so that
libmapserver.so is not linked against libfcgi. libfcgi
is now linked only against mapserv.

this patch closes #4296