Fix an issue with the complex polygon in oracle
The problem was when a polygon is composed from multiple line string and/or an arc polygon, the type of the shape is change for MS_SHAPE_LINE instead to keep the MS_SHAPE_POLYGON flag. So to solve it I validate the type provide by oracle and if it's a complex polygon I force to set the shape type to MS_SHAPE_POLYGON and the shape are now process as a polygon instead of a line string.