msResampleGDALToMap(): optimizations for tiled queries
If requesting at the raster resolution, on pixel boundaries,
and no reprojection is  involved, we don't need any resampling.
And if coordinates and dimensions match an integral subsampling
factor, no need to  add pixel margin.
This optimization helps a lot when operating with mode=tile
and that the underlying raster is tiled and share the same
tiling scheme as the queried tile mode.