msDrawRasterLayerGDAL(): find alpha band even if not 4th one (#5216)
Currently, in the absence of a BANDS processing directive,
msDrawRasterLayerGDAL() only checks that if the 4th band is the alpha
band when the raster has 4 or more bands. But for some products, the 4th
band might be the last one, so we'd better check the color interpretation
of all bands >= 4th one.