Add MapScript support for Python 3.x
I haven't had time to test this much, but this allows MapScript to
build and run against Python 3.4. The changes appear to continue to
work with Python 2.7.

If there are multiple Python versions installed on your machine,
the Python version that is selected is whatever CMake finds first
or the path set with -DPYTHON_EXECUTABLE=

Changes include:
* print ""                     ---> print("")
* raise E, "text"              ---> raise E("text")
* PyString_FromStringAndSize() ---> PyBytes_FromStringAndSize()
* Remove references to PyFile*