Output format: add a AGG/MIXED pseudo driver.
This driver redirects to a concrete driver when the image
has non-opaque pixels, and to another one when the image
is fully opaque.

Requires to define TRANSPARENT_FORMAT and OPAQUE_FORMAT
FORMAT_OPTION on OUTPUTFORMATs that instanciate that
driver. Those underlying formats can be AGG/PNG, AGG/JPEG
or a GDAL/ one (anyone that uses AGG/ for rendering)

Two predefined formats image/jpeg-png and image/jpeg-png8
are added.

Example of custom mixed formats:

    OUTPUTFORMAT
      NAME jpeg_low
      MIMETYPE "image/jpeg; quality=50"
      DRIVER AGG/JPEG
      IMAGEMODE RGB
      FORMATOPTION "QUALITY=50"
    END

    OUTPUTFORMAT
      NAME jpeg_low_png
      MIMETYPE "image/jpeg_low-png"
      DRIVER AGG/MIXED
      FORMATOPTION "TRANSPARENT_FORMAT=png"
      FORMATOPTION "OPAQUE_FORMAT=jpeg_low"
    END

    OUTPUTFORMAT
      NAME jpeg_high
      MIMETYPE "image/jpeg; quality=99"
      DRIVER AGG/JPEG
      IMAGEMODE RGB
      FORMATOPTION "QUALITY=99"
    END

    OUTPUTFORMAT
      NAME jpeg_high_png8
      MIMETYPE "image/jpeg_high-png8"
      DRIVER AGG/MIXED
      FORMATOPTION "TRANSPARENT_FORMAT=png8"
      FORMATOPTION "OPAQUE_FORMAT=jpeg_high"
    END