Fixed crash of mapserv due to memory corruption at large  values of variables of type double.
 I'll explain the reason for editing with mappostgis.c, starting at line 3640.

 In this case, for values node->tokenval.dblval> = 123456789012 (such values are real in the field),
strlen(snippet)> = 19,
and the memory is requested in the program only for 19 bytes. The unsolicited
memory is overwritten.

 It was possible without a check to request 318 bytes, this is enough to store number, when converting

to char [] having the longest string, - numeric_limits <double> :: max (),
but it is a pity to spend
memory when it is likely that most of the values are less of this number.

 Other edits are associated with preventing similar errors. Only there memory is sometimes requested
in other quantities.

 Developers should pay attention to other pieces of software, where it is used sprintf( ..., "%lf",...).