Add an offset to the scalebar
The scalebar is "stuck" to the edges of the map. This can make the
scalebar difficult to read when it is embedded in some mapping applications.

This PR adds the "OFFSET" key word to the scalebar object. The OFFSET X and Y
moves the scalebar away from the map a set number of pixels.