WCS: support providing creation options that are layer specific
If a Mapfile defines a GEOTIFF outputformat that is enabled in a WCS
layer, then it is possible to define a per-layer value for a creation option,
with a syntax like:

    "wcs_outputformat_GEOTIFF_creationoption_COMPRESS" "LZW"

to say that for 'GEOTIFF' output format, then COMPRESS=LZW should be passed to
the GDAL GTiff driver.

For GDAL 2.3.0 and the GRIB driver that supports band-specific creation options,
it is possible to define such band-specific creation options with for example

   "wcs_outputformat_GRIB_creationoption_BAND_1_IDS" "CENTER=54(Montreal) SUBCENTER=0 MASTER_TABLE=4 LOCAL_TABLE=0 SIGNF_REF_TIME=1(Start_of_Forecast) REF_TIME=2017-01-05T00:00:00Z PROD_STATUS=0(Operational) TYPE=2(Analysis_and_forecast)"

In case of band selection, the code will automatically renumber the band number.