Add additional EXPRESSION msautotests (#5897)
This pull request adds some more complex EXPRESSION tests to the msautotest suite. Focus is on combined expressions (AND, OR etc.) using the various supported syntaxes, several of which are not covered by tests.

Test results are CSV exports so easy to review and add new tests on different OSs.