gml_[item]_type: modify Date semantics, and add Time and DateTime
Currently, for vector features, we have a Date datatype that is used
indifferently for Date, Time, DateTime, which can cause approximations
in GML or other output formats (such as OGR GeoJSON), since the output
doesn't really know the exact input.
This change modifies the semantics of Date to mean date, without time,
and add Time (time, without date) and DateTime. The type autodetection
is modified in the MSSQL2008, OGR, Oracle and PostGIS backends. On the
output side, the WFS/GML and OGR output are modified to take into account
those 3 separate types.