WMS Client - Fixes GetLegendGraphic in WMS 1.3.0
SLD_VERSION parameter is mandatory in GetLegendGraphic as specified by SLD version 1.1.0. This commit adds this parameter when requesting WMS services with version 1.3.0 or higher.

In addition to this it also fixes style parameter: in GetLegendGraphic it should be named STYLE and not STYLES.

Two new testcases has been added for the WMS client which will test GetLegendGraphic functionality for both WMS 1.1.1 and WMS 1.3.0.