MVT: Improves backwards compatibility
This commit improves the backwards compatibility of MVT media type so that application/x-protobuf is acknowledged as well (e.g. it will still be advertised in GetCapabilities response).
In addition to this the commit also fixes the failing test case as well. It also adds a new testcase where the new media type is tested as well.