WCS netCDF output: allow generation of a 3D dataset from multiple bands
If the input dataset has multiple bands, it is now possible to generate
a 3D dataset with the extra dimension being typically time or Z.

This requires the following metadata item to be defined in
LAYER.METADATA:

"wcs_outputformat_{outputformat}_mdi_BAND_{bandnumber}_default_NETCDF_DIM_{dimname}" "{value}"

where:
- {outputformat} is substituted with the OUTPUTFORMAT.NAME value (typically "netCDF")
- {bandnumber} is substituted with the band number (between 1 and number of bands)
- {dimnname} is substituted with the name of the extra dimension. e.g time.
- {value} is substituted with the value taken on that extra dimension by the band. It must
  be a 32 bit integer, 64 bit integer or floating-point value.

'mdi' stands for metadata item
'default' indicates that this is for the GDAL default metadata item
NETCDF_DIM_ is a specific prefix used by the GDAL netCDF driver to
distinguish metadata items that define extra dimensions.

In addition to that, netCDF attributes can be defined on the extra
dimension by setting metadata items of the form

"wcs_outputformat_{outputformat}_mdi_default_{dimname}#{attrname}" "{value}"

where {attrname} is substituted with the name of the attribute to set on
the extra dimension (e.g. "standard_name", "units", etc.)

Extract from an example mapfile:

```
   # Metadata items to apply to the "time" extra dimension.
   "wcs_outputformat_netCDF_mdi_default_time#calendar" "standard"
   "wcs_outputformat_netCDF_mdi_default_time#standard_name" "time"
   "wcs_outputformat_netCDF_mdi_default_time#units" "hours since 1900-01-01 00:00:00"

   # Values of each band along the "time" extra dimension.
   "wcs_outputformat_netCDF_mdi_BAND_1_default_NETCDF_DIM_time" "10"
   "wcs_outputformat_netCDF_mdi_BAND_2_default_NETCDF_DIM_time" "20"
   "wcs_outputformat_netCDF_mdi_BAND_3_default_NETCDF_DIM_time" "30"
   "wcs_outputformat_netCDF_mdi_BAND_4_default_NETCDF_DIM_time" "40"
   "wcs_outputformat_netCDF_mdi_BAND_5_default_NETCDF_DIM_time" "50"
   "wcs_outputformat_netCDF_mdi_BAND_6_default_NETCDF_DIM_time" "60"
   "wcs_outputformat_netCDF_mdi_BAND_7_default_NETCDF_DIM_time" "70"
   "wcs_outputformat_netCDF_mdi_BAND_8_default_NETCDF_DIM_time" "80"
   "wcs_outputformat_netCDF_mdi_BAND_9_default_NETCDF_DIM_time" "90"
```