Refactor mapcairo size extraction logic
The matrix of conditionals around SVG size extraction has been refactored in mapcairo.c. With this change, the symbols' height and width are better handled depending on the presence of width and height attributes, their units, and the viewBox. The conditions now account for newest LIBRSVG versions and fallback for older ones.