Add 4 new COMPOSITE.COMPOP blending operations
Cf https://pycairo.readthedocs.io/en/latest/reference/enums.html#cairo.Operator.HSL_HUE

- "hsl-hue": Creates a color with the hue of the source and the saturation
  and luminosity of the target.
- "hsl-luminosity": Creates a color with the luminosity of the source
  and the hue and saturation of the target. This produces an inverse
  effect to hsl-color
- "hsl-saturation": Creates a color with the saturation of the source and the hue and
  luminosity of the target. Painting with this mode onto a gray area produces no change.
- "hsl-color": Creates a color with the hue and saturation of the source and the
  luminosity of the target. This preserves the gray levels of the target and is
  useful for coloring monochrome images or tinting color images.

That last mode is typical the most useful, for example when draping a
colorized terrain onto a hillshade layer, or to colorize a greyed LIDAR
shaded surface model with a green mask of a canopy model.

~~~

Implemented in the Cairo backend using the corresponding Cairo operators
(which ultimately go to using Pixman), or in the AGG backend, by
manually extracting the related Pixman code and adapting it to AGG (or
when building WITH_PIXMAN=ON, bu using the Pixman operation)

Note: if using OPACITY < 100 with the Cairo backend (or AGG
WITH_PIXMAN=ON), there's a bug in current Pixman versions that result in
wrong colors. Cf https://gitlab.freedesktop.org/pixman/pixman/-/merge_requests/97
The adaptation in the AGG backend isn't affected by this bug.