Fix style scale factor inheritance
Commit ede4953 introduced SIZEUNITS inheritance layer->class->style/label.
When LAYER STYLEITEM is in use, for each feature msUpdateStyleFromString
is called but no attempts are made to ensure correct scale factor values
for newly created style item thus all features get scale factor 1.0,
that might be different from value it should have. This wasn't a problem
before ede4953 as layer scale factor was always used to render features
with STYLEITEM.

This commit introduces a workaround by lifting out code from the only
place where scale factor is adjusted in a way that it can be reused also
in the STYLEITEM case.