SLD Filters missing due to logicial keywords in expressions (#7178)
When using WMS GetStyles requests to generate SLD I noticed some filters were missing. On debugging it became apparent that any string containing a logical operator was failing to generate a filter:

```
EXPRESSION ("[Property1]" = "Not Applicable")
```

Was missing the expected filter:

```xml
<ogc:Filter><ogc:PropertyIsEqualTo><ogc:PropertyName>Property1</ogc:PropertyName><ogc:Literal>Not Applicable</ogc:Literal></ogc:PropertyIsEqualTo></ogc:Filter>
```

`msSLDNumberOfLogicalOperators` had a simple search for `AND`, `NOT`, and `OR` that was finding these words in strings causing errors. This fix attempts to tokenize the string, leaving quoted strings intact to avoid this. It is by no means perfect, and I also had to change previous test cases to have spaces around keywords e.g. `(([FID] > 1)And([FID] <= 10))` so that they pass. 

MapServer seems to be missing a parsing engine for expressions (or at least I can't find anything in the codebase). Without this any SLD expression building is going to be flaky. See also #5894.