Allow <> to be used as a logical operator in expressions (#7182)
This pull request allows <> to be used for "not equal to" logical expressions, along with the already supported ne and !=. Fixes #6950.

When creating SLD filters expressions are created by checking values in strings, so modifications have been added to mapogcsld.cpp have been made to support these. There are still issues around this approach for creating SLD filters as noted in #7178.