Add HTTP status code for WMS requests, if MS_WMS_ERROR_STATUS_CODE config option/env variable are set to ON
Currently when one issues an HTTP request that ends up in error,
MapServer doesn't emit a HTTP Status header response. Hence the
webserver will default to generate a 200 OK one.

This is an historical behavior used by other WMS server implementations
(GeoServer, Degree) due to the WMS specification being unclear on what
do do. Given that WMS is in a retirement phase, it would be risky to
change that behavior by default.

That said there are WMS clients that would enjoy having a proper HTTP
400 or 500 status code. Hence if the MS_WMS_ERROR_STATUS_CODE
environment variable / config option is set to ON (typically in the ENV
section of the mapserver configuration file), MapServer will emit those
status codes.

This relies on code to call the msSetErrorWithStatus() new function. A
pass has been done in mapwms.cpp to call it. And a more moderate one in
a few places of mapfile.c and mapservutil.c related to generic mapfile
handling.