Update Windows CI and Python Versions (#7171)
As of 2024-10-07 Python 3.8 is no longer supported. The new 3.13 Python was released on 2024-10-07.
This pull request removes 3.8 and adds 3.13. The cache folder for SWIG was also updated to match the version in use.
See https://www.python.org/downloads/ for details.

Other changes:
- Image updated to Visual Studio 2022
- SWIG updated to 4.3
- SDK updated to https://www.gisinternals.com/query.html?content=filelist&file=release-1930-x64-dev.zip

There is a sqlite3.dll incompatibility between Python 3.13 (version 3.45.3) and the SDK (version 3.47.1). The CI copies the DLL into the Python installation.