UVRaster: reproject wind vector from layer projection to map/request projection
Before this PR, the angle of the wind vector wasn't modified if
reprojection was involved. With this PR, we now build a small vector
representing the wind direction in the layer projection and reproject it
to the target projection. We use that to correct the vector direction,
while preserving the original vector length.
The newly added test case shows a raster in EPSG:4326 that has a pure
west-oriented wind, and queried in a north polar stereographic
projection, center at the north pole. The clock wise oriented
arrows along circles are the expected result.